+++ ask-smapi-model/ask_smapi_model/v1/skill/account_linking/__init__.py
@@ -15,6 +15,7 @@
 from __future__ import absolute_import
 
 from .account_linking_platform_authorization_url import AccountLinkingPlatformAuthorizationUrl
+from .account_linking_request_payload import AccountLinkingRequestPayload
 from .account_linking_response import AccountLinkingResponse
 from .access_token_scheme_type import AccessTokenSchemeType
 from .account_linking_request import AccountLinkingRequest